git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@577 235a714a-1e17-0410-b3bb-947d69a96898

+++ include/mw/common/Event.h
@@ -8,7 +8,7 @@ namespace famouso {
 	namespace mw {
 
 struct Event {
-  typedef uint32_t Type;
+  typedef uint16_t Type;
   const Subject &subject;
   Type length;
   uint8_t *data;
